Initialize Courier repo.
Summary:
This includes:
* .gitignore
* Arc files
* Basic SBT files (e.g. project/Build.scala)
* Basic OSS (LICENSE, CONTRIBUTING.md)
* Coursera artifactory

One of the big open issues here is how to publish to both Coursera artifactory (for rapid iteration) and to maven central (for wide distribution). For those that have not worked with maven central, publishing is annoyingly manual process and there is a delay of 4-8 hours(?) before a published artifact becomes available. Some options:

1. Keep both maven central and Coursera artifactory settings in Build.scala, add a SBT settings to toggle which one is in use.
2. Use bittray to manage publishing. Need to test it out more to see if it can do a reasonable job of this.
3. Find a way to inject repo settings into the build, only include maven central settings here, but inject in the Coursera artifactory settings for our Jenkins builds that publish to artifactory.

Test Plan:
* `sbt "projects"`
* `arc lint`

Reviewers: saeta

Reviewed By: saeta

Subscribers: nick

Differential Revision: https://phabricator.dkandu.me/D15872