Hook courier into the rest.li-sbt-plugin's extension point.
Summary:
Earlier today I made a few changes to https://github.com/coursera/rest.li-sbt-plugin/commits/master. I also published the changes to maven central as `org.coursera.courier:courier-sbt-plugin_2.10:0.0.1`. This needs to be Scala 2.10 because SBT 13 is on scala 2.10. If we are able to get these changes merged into the official linkedin/rest.li-sbt-plugin master branch, then we will not need to publish our own version to maven central, but for now we do.

In the sbt plugin, I've added an extension point for the Courier scala generator. It will be used like so:

```
.settings(restliPegasusGeneratorClassname := classOf[ScalaDataTemplateGenerator].getName)
```

This diph adds the `ScalaDataTemplateGenerator` class, which contains all the boilerplate needed to integrate with the pegasus data generator libraries, and connects it to a not-yet-implemented treehugger generator.

Test Plan: * Manually tested that this can be published and consumed by a SBT project and that the generator is run when it should be.

Reviewers: saeta

Reviewed By: saeta

Subscribers: nick