Add support for Fixed, add generateTyperefs option
Summary:
For completeness, I've added support for the "Fixed" type. See avro docs for details. I don't expect us to need this, but it's easy enough to add, so I've put it in for completeness.

I've also added a flag to allow us to avoid generating `TyperefInfo` classes. We don't need them initially, and they can be a bit annoying because, if you, say, have a typeref to a hand written class named `CourseId` and the pegasus typeref schema is also named `CourseId`, then it will generate an additional `CourseId` class (that extends `TyperefInfo`).  This can be annoying for developers when selecting which class to import in an IDE because they would then see two `CourseId` types and might not know which one is the right one to pick (particularly if their packages have similar names).

All a `TyperefInfo` class provides is easy access to the typeref's schema for cases where the typeref is used alone.  The schema is also inlined in schemas the typeref is used in, and so TyperefInfos are not required for any of our current needs.  If we ever need TyperefInfos (and we might at some point in Naptime to support, say, using pegasus coercers for serialization of a typeref query param value, or to apply custom validation rules defined in a typeref to a query param value) then we can start generating them at that time (and maybe sort out a way to generate them so they are clearly distinct from the types they reference).

Lastly, I've removed validation from `DataTemplates.makeImmutable`. We will instead make it policy for validation to be run at system boundaries (Courral already does this).  This is mainly to ensure that systems make explicit decisions about validation.  Each system should explicitly configure validation with appropriate settings for the data and codec(s) it is using, and should message validation errors in an appropriate way. There are also performance considerations: Validation has a non-negligible cost, so should be performed when and where it is needed (usually at system boundaries or when accessing data stores or caches). Also since validation fix-up requires making a full copy of the data when data is set as ReadOnly, it's best that we carefully manage when it is called.  We can most easily do this at system boundaries which, when deserializing data, should have exclusive access to it and should be able to validate and fix it up before marking it as ReadOnly.

Test Plan: Added unit tests.

Reviewers: saeta, danchia

Reviewed By: danchia

Subscribers: nick