Add support for typed map keys
Summary:
Per https://docs.google.com/document/d/1KGlwLe1IBhy32uBYKmELM8c4_9WCwJdo_y5Z8MSkDJg/edit#heading=h.9i0fl5ks7968, this adds support to Courier for typed map keys.

This introduces the first major schema language feature not provided by Pegasus. As a result, the integration with the generator utilities that pegasus provides, and that Courier has been able to use successfully up to this point, is not working out nearly as well for this change.  The only viable approach I've found, in the short term, is to duplicate two classes from pegasus and modify them to suit our purposes. I will continue to work with the Rest.li team to add extension points to the the generator utility classes, mainly `TemplateSpecGenerator`, so that in the long term, we can delete these duplicated classes.  Because we do not intended to keep these classes around permenently, I have not ported them to Scala.  If any reviewers have suggestions on how to do this integration in a cleaner way, please let me know, I'm unhappy to have to duplicate these classes, but have been unable to find a cleaner solution that allows us to make progress on this feature.

I've also fixed Courier to use "scala" instead of "java" for custom types, as it should.

Test Plan:
* Added unit tests
* Locally tested with infra-service and the notification preferences models that need typed map keys

Reviewers: saeta, danchia

Reviewed By: danchia

Subscribers: nick

Differential Revision: https://phabricator.dkandu.me/D19258