escape class fields reserved by RecordTemplate and Product
Summary: @danchia discovered that it's possible for Courier to generate a Scala class that cannot compile if a reserved class field name like `data`, `schema`, `copy` or `hashCode` is used in as a pegasus record field name.  This adds an escaping rule to avoid these collisions.

Test Plan: Added unit test

Reviewers: josh, zhaojun, danchia

Reviewed By: danchia

Subscribers: danchia

Differential Revision: https://phabricator.dkandu.me/D21204