Remove dependency on StrictLogging from courier-generator
Summary: Unfortunately, type-safe has introduce a new version of StrictLogging that is binary incompatible with the previous one.  Since Courier is opensource, and developers may be using other SBT plugins running different versions of StrictLogging, it's better for our developers if we simply remove it, which fortunately, is trivial as we are not using it for anything.

Test Plan: * sbt fulltest

Reviewers: danchia

Reviewed By: danchia

Subscribers: saeta

Differential Revision: https://phabricator.dkandu.me/D21424