Remove dates from generated scala files.
Summary:
In infra-services, I've noticed, as the number of schemas we have has grown, that SBT seems to be unnecessarily recompiling all the scala files generated by courier whenever any of them are changed. This is already unacceptably slow.

Since SBT checks hashes to determine what to recompile, the obvious fix is to remove the timestamps that Courier was previously including in the @Generated annotation of each generated scala file.

With this fix in place, small changes to schemas only take 3-7 seconds for SBT to re-generate and incrementally compile.

Test Plan:
Test this in infra-services, verify that a change to a single unreferenced schema results in only one scala class being recompiled:

```
> compile
[info] Courier: Generating Scala bindings for .pdsc and .courier files.
[info] Compiling 1 Scala source to /Users/jbetz/base/coursera/infra-services/libs/models/target/scala-2.11/classes...
```