Fix Enum Symbol Docs (#29)
Enum Symbol Docs should be set not on properties but directly on
setSymbolDocs()

Additionally, in order to ensure the schema parsing & serialization is lossless,
we add a round-trip test where we construct a schema, serialize it,
and re-parse it. We then ensure that the objects are equal to each
other.