Removed unneeded imports. Checked-in module config
Why check in the module configuration? It turns out that IDEA does a bad job
of importing plugin modules from source. In particular, it fails to recognize
them as plugin modules at all, thinking instead that they're just plain Java.
The consequence of this mis-identification is that the build configurations get
messed up, making it impossible to build your plugin.  The manual procedure to
fix this mis-identification is cryptic, so there's a clear benefit to checking
in the project files.

Why would you not check in the module files?
  * Someone using a different IDE (for developping IDEA plugins???)
  * IDK?